+++ src/Credfeto.Aur.Mirror.Server/Services/AurRepos.cs
@@ -78,6 +78,11 @@ public sealed class AurRepos : IAurRepos
         }
     }
 
+    public string GetRepoBasePath(string repoName)
+    {
+        return Path.Combine(path1: this._serverConfig.Storage.Repos, $"{repoName}.git");
+    }
+
     private async ValueTask<byte[]> RequestPackagesUpstreamAsync(
         ProductInfoHeaderValue? userAgent,
         CancellationToken cancellationToken